+++ src/components/RichTextEditor/types.ts
@@ -1,7 +1,7 @@
 /* (c) Copyright Frontify Ltd., all rights reserved. */
 
 import { CSSProperties, ReactElement } from 'react';
-import { ListStyles, TextStyles } from './EditorActions/Plugins/TextStylePlugin/TextStyleDropdown/types';
+import { TextStyles } from './EditorActions/Plugins/TextStylePlugin/TextStyles';
 import { EditorActions } from './utils/actions';
 
 export type DesignTokens = Partial<Record<TextStyles, CSSProperties>>;
@@ -23,8 +23,6 @@ export type ButtonGroupProps = {
 };
 
 export type ButtonGroupWidths = { actions: EditorActions[]; buttonGroupWidth: number; index: number }[];
-
-export type AvailableTextStyles = ListStyles & TextStyles;
 
 export enum PaddingSizes {
     None = 'tw-p-0',
